fix: 중복 및 데드 파라미터 제거
