+++ features/auth/fetchers.ts
@@ -1,0 +1,5 @@
+import { apiRequest } from '@/shared/lib/api';
+
+import type { UserInfoResDto } from './types';
+
+export const getMyProfile = (): Promise<UserInfoResDto> => apiRequest<UserInfoResDto>('/users/profile');
