+++ features/auth/kakao/types.ts
@@ -2,12 +2,3 @@ export interface LoginInfoResDto {
   accessToken: string;
   userId: number;
 }
-
-export interface UserInfoResDto {
-  userId: number;
-  username: string | null;
-  email: string;
-  bio: string | null;
-  profileImageUrl: string | null;
-  userType: string;
-}
